@@ -1017,7 +1017,7 @@ fn set_function_name(
             )
         }
         PropertyKey::String(string) => string.clone(),
-        PropertyKey::Index(index) => js_string!(format!("{}", index)),
+        PropertyKey::Index(index) => js_string!(format!("{index}")),
     };
 
     // 3. Else if name is a Private Name, then
